requirements.txt
kivy
pyjnius